Move attachment registry to tooling crate, alongside tool registry
