+++ crates/assistant_tooling/src/assistant_tooling.rs
@@ -1,6 +1,8 @@
 mod assistant_context;
+mod attachment_registry;
 mod tool_registry;
 
+pub use attachment_registry::{AttachmentRegistry, LanguageModelAttachment, UserAttachment};
 pub use tool_registry::{
     LanguageModelTool, ToolFunctionCall, ToolFunctionDefinition, ToolOutput, ToolRegistry,
 };
